requirements.txt
imagehash
matplotlib
numpy
pandas
Pillow
scikit-image
